+++ src/components/Chat/index.ts
@@ -1,1 +1,2 @@
 export * from "./ChatContent";
+export * from "./ChatField";