+++ src/components/Chat/stories/ChatField.stories.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { Meta, StoryFn } from "@storybook/react";
+
+import { ChatField } from "../../../index";
+
+export default {
+    title: "Components/Chat/ChatField",
+    component: ChatField,
+    argTypes: {},
+} as Meta<typeof ChatField>;
+
+let forceupdate = 0;
+const TemplateFull: StoryFn<typeof ChatField> = (args) => <ChatField {...args} key={forceupdate++} />;
+
+export const Default = TemplateFull.bind({});
+Default.args = {
+    onSubmit: (value) => alert(value),
+};